lavc/vvc: Change intra pixel type to `int16_t`
Previously the pixels were stored as `int`s. The maximum bit depth
supported by VVC is 16, hence an `int16_t` is guaranteed to be
sufficient. Smaller sizes allow for more pixels to be packed into an
SIMD register and therefore faster assembly optimisations. Further
optimisation could be achieved by using `int8_t`s if the bit depth of
the current video is only 8.

tests/checkasm/vvc_itx: Use `int16_t` dst arrays

tests/checkasm/vvc_itx: Fix dst stride

+++ libavcodec/vvc/vvc_intra_template.c
@@ -429,7 +429,7 @@ static int FUNC(lmcs_derive_chroma_scale)(VVCLocalContext *lc, const int x0, con
 }
 
 // 8.7.5.3 Picture reconstruction with luma dependent chroma residual scaling process for chroma samples
-static void FUNC(lmcs_scale_chroma)(VVCLocalContext *lc, int *dst, const int *coeff,
+static void FUNC(lmcs_scale_chroma)(VVCLocalContext *lc, int16_t *dst, const int16_t *coeff,
     const int width, const int height, const int x0_cu, const int y0_cu)
 {
     const int chroma_scale = FUNC(lmcs_derive_chroma_scale)(lc, x0_cu, y0_cu);
